[Uploader] Move files to dedicated folder
