+++ app/javascript/src/javascripts/uploader/checkbox.vue
