+++ app/javascript/src/javascripts/uploader/file_preview.vue
