+++ app/javascript/src/javascripts/uploader/source.vue
