+++ app/javascript/src/javascripts/uploader/tag_preview.vue
